Устранил баг с мапперами, добавил двусторннюю связь между User, Ads, AdsComment
